refactor: Marketplace `AppsPageContentBody` structure
